backend/requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django==3.2.18
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.1.1
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==1.7.0
iniconfig==2.0.0
isort==5.11.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.0
pep8-naming==0.13.3
Pillow==9.4.0
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
python3-openid==3.2.0
pytz==2023.2
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
zipp==3.15.0
